[WIP] Feature/toolkit go (#385)
* base toolkit package

* base revisions & structure for Go toolkit

* basics - downloading Go archive

* extension - different for Windows

* Go toolkit : retry download & base install (uncompress)

* Bootstrap implemented for Go toolkit

sets the required PATH and GOROOT

* checkAndInstallStepDependencies moved to a separate function

* executeStep moved to a separate function - will be the base point for Toolkit Run

* debug option to use System (installed) Tools instead of BITRISE_HOME ones

* Debug option logging change

* retry Wait - fix

* go version - moved to top

and download location extension fix

* setup / dependency install : error passing fix

not the right `error` was returned in "Failed to install plugin, error"
and a bit of log format revision, for better progress indicators

* base Bash and Go toolkit - WIP

it can now handle it, print it, etc. but still WIP, e.g. for Go it uses `go run` right now instead of a proper `go build` and then a simple "run"

* one more getRunningStepHeaderSubSection revision

* one more rev in step header printer

* base Toolkit check & setup in `bitrise setup`

* finishing Go toolkit Install and Check

can now properly detect if a system installed Go is available and sufficient, or if that's not available install the toolkit

* toolkit : PrepareForStepRun added

* finishing Go Toolkit handing - basics are all done!!