toolkit - new method IsToolAvailableInPATH
Check whether the toolkit's tool (e.g. Go, Ruby, Bash, ...) is available in the current PATH

This way Bitrise CLI can skip the toolkit's Bootstrap method if there's a system installed version of the tool.

Instead of requiring the toolkit to properly implement Bootstrap and follow this guideline in the docs this is now handled in a central code.