deps install revision - use both "which" and package manager
First use "which" to check whether the tool is available, which can be available from another source, outside of the package manager (e.g. use NVM to install Node.js versions).

If which would not find the binary, fallback to check with the package manager (brew/apt) whether the package is installed. This is useful mainly if someone forgets to specify the BinName for a package where the binary's name does not match the package name - e.g. "awscli" is the package, for which, even if it's installed, "which awscli" will never find the binary, as the binary is "aws". In this case check the package (awscli in this example) directly with the package manager too, to see if the package with that name is installed.